[JIT] Clone runOptimizations and similar functions for profiling executor.
Thing change will allow us to more freely experiment with pass pipelines
in the profiling executor without affecting passes in the legacy
executor. Also, it somewhat helps to keep all passes in one place to be
able to tell what's going on.

Currently this change should not affect any behavior as I copied the
passes exactly as they've been invoked before, but we will probably want
to change these pipelines in a near future.

ghstack-source-id: 60872735a36d4cb99339b2cc11f91ffc1749812f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42656