Restrict conversion to SmallVector
The old implementation allowed calling SmallVector constructor and operator= for any type without restrictions,
but then failed with a compiler error when the type wasn't a collection.

Instead, we should only use it if Container follows a container concept and just not match the constructor otherwise.

This fixes an issue @kimishpatel was running into.

Differential Revision: [D22983020](https://our.internmc.facebook.com/intern/diff/D22983020/)

ghstack-source-id: 109370513
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42694